re-organize tests in rlp_circuit_fsm into its own dir
